third_party/blink/web_tests/platform/generic/css3/blending/background-blend-mode-property-expected.txt

Test the parsing of the background-blend-mode property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('background-blend-mode') is 'normal, normal'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('background-blend-mode') is 'multiply, multiply'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('background-blend-mode') is 'screen, screen'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('background-blend-mode') is 'overlay, overlay'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('background-blend-mode') is 'darken, darken'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('background-blend-mode') is 'lighten, lighten'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('background-blend-mode') is 'color-dodge, color-dodge'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('background-blend-mode') is 'color-burn, color-burn'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('background-blend-mode') is 'hard-light, hard-light'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('background-blend-mode') is 'soft-light, soft-light'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('background-blend-mode') is 'difference, difference'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('background-blend-mode') is 'exclusion, exclusion'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('background-blend-mode') is 'hue, hue'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('background-blend-mode') is 'saturation, saturation'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('background-blend-mode') is 'color, color'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('background-blend-mode') is 'luminosity, luminosity'
PASS successfullyParsed is true

TEST COMPLETE
